+++ lib/simple_form/components/labels.rb
@@ -38,10 +38,14 @@ module SimpleForm
       end
 
       def label_html_options
-        label_html_classes = SimpleForm.generate_additional_classes_for.include?(:label) ? [input_type, required_class, SimpleForm.label_class].compact : []
+        label_html_classes = SimpleForm.additional_classes_for(:label) {
+          [input_type, required_class, SimpleForm.label_class].compact
+        }
 
         label_options = html_options_for(:label, label_html_classes)
-        label_options[:for] = options[:input_html][:id] if options.key?(:input_html) && options[:input_html].key?(:id)
+        if options.key?(:input_html) && options[:input_html].key?(:id)
+          label_options[:for] = options[:input_html][:id]
+        end
         label_options
       end