+++ lib/simple_form/wrappers/root.rb
@@ -24,7 +24,7 @@ module SimpleForm
 
       def html_classes(input, options)
         css = options[:wrapper_class] ? Array.wrap(options[:wrapper_class]) : @defaults[:class]
-        css += input.html_classes if SimpleForm.generate_additional_classes_for.include?(:wrapper)
+        css += SimpleForm.additional_classes_for(:wrapper) { input.html_classes }
         css << (options[:wrapper_error_class] || @defaults[:error_class]) if input.has_errors?
         css << (options[:wrapper_hint_class] || @defaults[:hint_class]) if input.has_hint?
         css